small fixes

+++ src/Rocket.cpp
@@ -6,7 +6,7 @@ Rocket::Rocket(glm::vec2 position, glm::vec2 direction){
 }
 
 int Rocket::intersectsAsteroid(std::vector<Asteroid*> asteroids){
-  for(int i=0;i<asteroids.size();i++){
+  for(size_t i=0; i<asteroids.size(); i++){
     std::vector<glm::vec2> BB = asteroids[i]->BB;
     glm::vec2 asteroid_position = asteroids[i]->center;
     if((this->position.x < (asteroid_position.x + BB[0].x)) && (this->position.x > (asteroid_position.x + BB[1].x)) && (this->position.y < (asteroid_position.y + BB[2].y)) && (this->position.y >(asteroid_position.y + BB[0].y))){
